refactor: organize social media
