+++ src/App.js
@@ -2,7 +2,7 @@ import './App.css';
 import Description from './components/Description';
 import Footer from './components/Footer';
 import SocialLinks from './components/containers/domains/socials/SocialLinks';
-import SocialMediaLinks from './components/SocialMediaLinks';
+import SocialMediaLinks from './components/containers/domains/social-media/SocialMediaLinks';
 import SubTitle from './components/SubTitle';
 import Title from './components/Title';
