+++ src/components/containers/domains/social-media/SocialMedia.js
@@ -1,0 +1,20 @@
+
+export default function SocialMedia() {
+    return [
+        {
+            text: "YouTube",
+            hoverText: "My Profile on YouTube",
+            url: "https://www.youtube.com/channel/UC19Ueh3p2qYWkWUF8qFz0Pg",
+        },
+        {
+            text: "Twitter",
+            hoverText: "My Profile on Twitter",
+            url: "https://twitter.com/Chrimle",
+        },
+        {
+            text: "Instagram",
+            hoverText: "My Profile on Instagram",
+            url: "https://www.instagram.com/chrimle/",
+        },
+    ];
+}
