+++ src/components/containers/domains/social-media/SocialMediaLinks.js
@@ -1,13 +1,7 @@
 import { View } from 'react-native';
-import TextLinkButton from './buttons/TextLinkButton';
-import theme from '../theme';
-
-const socialMediaLinks = [
-  { text: "YouTube", hoverText: "My Profile on YouTube", url: "https://www.youtube.com/channel/UC19Ueh3p2qYWkWUF8qFz0Pg" },
-  { text: "Twitter", hoverText: "My Profile on Twitter", url: "https://twitter.com/Chrimle" },
-  { text: "Instagram", hoverText: "My Profile on Instagram", url: "https://www.instagram.com/chrimle/" },
-];
-
+import TextLinkButton from '../../../buttons/TextLinkButton';
+import theme from '../../../../theme';
+import SocialMedia from './SocialMedia';
 
 export default function SocialMediaLinks() {
 
@@ -19,7 +13,7 @@ export default function SocialMediaLinks() {
     </text>
 
     <View style={{ flexDirection: 'row' }}>
-      {socialMediaLinks.map(socialMediaLink =>
+      {SocialMedia().map(socialMediaLink =>
         <TextLinkButton text={socialMediaLink.text} url={socialMediaLink.url} hoverText={socialMediaLink.hoverText} />
       )}
     </View>
